Specify action taken when exceeded on a per-worker basis
You can now pass a proc in to ann exceeded option to the throttler options.

If this is passed in, then the default behavior (which is to perform_in(delay)
will be used instead.

The proc that is passed in is instance_exec'd which gives access to all the
rate limiters variables.

This is probably not that safe...