Can now specify the behavior to occur when a rate limit is exceeded.
The behavior can be specified with the :exceeded property in throttler
options for the class.

By leaving this nil or setting it to :retry, the previous behavior of
retrying the job occurrs.

To specify different behavior, set the :exceeded option to a Proc that
takes up to 4 arguments:

1) The period/delay of the rate limiter
2) The worker object that is processing the job.
3) The payload (job arguments)
4) The queue the job came from.

A lambda can also be used if all 4 of the above arguments are specified
in the lambda definition.